Project finished

+++ src/main/java/com/workintech/s18d1/dao/BurgerDao.java
@@ -10,11 +10,11 @@ public interface BurgerDao {
 
     Burger save(Burger burger);
 
-    Burger findById(int id);
+    Burger findById(long id);
 
     List<Burger> findAll();
 
-    List<Burger> findByPrice(double price);
+    List<Burger> findByPrice(int price);
 
     List<Burger> findByBreadType(BreadType breadType);
